+++ Animate/components/com_render_specular.ts
