+++ core/systems/sys_camera.ts
@@ -1,4 +1,4 @@
-import {multiply, perspective} from "../../common/mat4.js";
+import {get_translation, multiply, perspective} from "../../common/mat4.js";
 import {Entity, Game} from "../game.js";
 import {Has} from "../world.js";
 
@@ -39,4 +39,5 @@ function update(game: Game, entity: Entity) {
     }
 
     multiply(camera.Pv, camera.Projection, transform.Self);
+    get_translation(camera.Position, transform.World);
 }
