+++ core/systems/sys_render.ts
@@ -1,8 +1,10 @@
 import {Material} from "../../common/material.js";
 import {GL_COLOR_BUFFER_BIT, GL_DEPTH_BUFFER_BIT, GL_UNSIGNED_SHORT} from "../../common/webgl.js";
 import {DiffuseLayout} from "../../materials/layout_diffuse.js";
+import {SpecularLayout} from "../../materials/layout_specular.js";
 import {RenderKind} from "../components/com_render.js";
 import {RenderDiffuse} from "../components/com_render_diffuse.js";
+import {RenderSpecular} from "../components/com_render_specular.js";
 import {Transform} from "../components/com_transform.js";
 import {Game} from "../game.js";
 import {Has} from "../world.js";
@@ -30,6 +32,9 @@ export function sys_render(game: Game, delta: number) {
                     case RenderKind.Diffuse:
                         use_diffuse(game, render.Material);
                         break;
+                    case RenderKind.Specular:
+                        use_specular(game, render.Material);
+                        break;
                 }
             }
 
@@ -41,6 +46,9 @@ export function sys_render(game: Game, delta: number) {
             switch (render.Kind) {
                 case RenderKind.Diffuse:
                     draw_diffuse(game, transform, render);
+                    break;
+                case RenderKind.Specular:
+                    draw_specular(game, transform, render);
                     break;
             }
         }
@@ -58,6 +66,25 @@ function draw_diffuse(game: Game, transform: Transform, render: RenderDiffuse) {
     game.Gl.uniformMatrix4fv(render.Material.Locations.World, false, transform.World);
     game.Gl.uniformMatrix4fv(render.Material.Locations.Self, false, transform.Self);
     game.Gl.uniform4fv(render.Material.Locations.Color, render.Color);
+    game.ExtVao.bindVertexArrayOES(render.Vao);
+    game.Gl.drawElements(render.Material.Mode, render.Mesh.IndexCount, GL_UNSIGNED_SHORT, 0);
+    game.ExtVao.bindVertexArrayOES(null);
+}
+
+function use_specular(game: Game, material: Material<SpecularLayout>) {
+    game.Gl.useProgram(material.Program);
+    game.Gl.uniformMatrix4fv(material.Locations.Pv, false, game.Camera!.Pv);
+    game.Gl.uniform3fv(material.Locations.Eye, game.Camera!.Position);
+    game.Gl.uniform4fv(material.Locations.LightPositions, game.LightPositions);
+    game.Gl.uniform4fv(material.Locations.LightDetails, game.LightDetails);
+}
+
+function draw_specular(game: Game, transform: Transform, render: RenderSpecular) {
+    game.Gl.uniformMatrix4fv(render.Material.Locations.World, false, transform.World);
+    game.Gl.uniformMatrix4fv(render.Material.Locations.Self, false, transform.Self);
+    game.Gl.uniform4fv(render.Material.Locations.ColorDiffuse, render.ColorDiffuse);
+    game.Gl.uniform4fv(render.Material.Locations.ColorSpecular, render.ColorSpecular);
+    game.Gl.uniform1f(render.Material.Locations.Shininess, render.Shininess);
     game.ExtVao.bindVertexArrayOES(render.Vao);
     game.Gl.drawElements(render.Material.Mode, render.Mesh.IndexCount, GL_UNSIGNED_SHORT, 0);
     game.ExtVao.bindVertexArrayOES(null);
